flask-backend/api/requirements.txt
PyJWT
PyMySQL
Flask
python-decouple
gunicorn
flask-cors
pytest
requests
